ENH: separate stoppers from valves, add new pps stopper class
