+++ pcdsdevices/valve.py
@@ -7,8 +7,8 @@ from enum import Enum
 from ophyd import Component as Cpt
 from ophyd import Device, EpicsSignal, EpicsSignalRO, EpicsSignalWithRBV
 
-from .inout import InOutPositioner, InOutPVStatePositioner
 from .interface import LightpathMixin
+from .stopper import PPSStopper, Stopper  # noqa import PPS for backcompat
 
 logger = logging.getLogger(__name__)
 
@@ -22,64 +22,6 @@ class Commands(Enum):
 class InterlockError(PermissionError):
     """Error when request is blocked by interlock logic."""
     pass
-
-
-class Stopper(InOutPVStatePositioner):
-    """
-    Controls Stopper.
-
-    A base class for a device with two limits switches controlled via an
-    external command PV. This fully encompasses the controls Stopper
-    installations as well as un-interlocked `GateValve` s.
-
-    Parameters
-    ----------
-    prefix : str
-        Full PPS Stopper PV.
-
-    name : str
-        Alias for the stopper.
-
-    Attributes
-    ----------
-    commands : ~enum.Enum
-        An enum with integer values for `~Commands.open_valve` and
-        `~Commands.close_valve` values.
-    """
-
-    # Limit-based states
-    open_limit = Cpt(EpicsSignalRO, ':OPEN', kind='normal')
-    closed_limit = Cpt(EpicsSignalRO, ':CLOSE', kind='normal')
-
-    # Information on device control
-    command = Cpt(EpicsSignal, ':CMD', kind='omitted')
-    commands = Commands
-
-    _state_logic = {'open_limit': {0: 'defer',
-                                   1: 'OUT'},
-                    'closed_limit': {0: 'defer',
-                                     1: 'IN'}}
-    # QIcon for UX
-    _icon = 'fa.times-circle'
-
-    tab_whitelist = ['open', 'close']
-
-    def __init__(self, prefix, *, name, **kwargs):
-        super().__init__(prefix, name=name, **kwargs)
-
-    def _do_move(self, state):
-        if state.name == 'IN':
-            self.command.put(self.commands.close_valve.value)
-        elif state.name == 'OUT':
-            self.command.put(self.commands.open_valve.value)
-
-    def open(self, **kwargs):
-        """Open the stopper."""
-        return self.remove(**kwargs)
-
-    def close(self, **kwargs):
-        """Close the stopper."""
-        return self.insert(**kwargs)
 
 
 class GateValve(Stopper):
@@ -100,6 +42,7 @@ class GateValve(Stopper):
 
     # Commands and Interlock information
     command = Cpt(EpicsSignal, ':OPN_SW', kind='omitted')
+    commands = Commands
     interlock = Cpt(EpicsSignalRO, ':OPN_OK', kind='normal')
 
     # QIcon for UX
@@ -121,49 +64,6 @@ class GateValve(Stopper):
         opening.
         """
         return not bool(self.interlock.get())
-
-
-class PPSStopper(InOutPositioner):
-    """
-    PPS Stopper.
-
-    Control of this device only available to PPS systems. This class merely
-    interprets the summary of limit switches to let the controls system know
-    the current position.
-
-    Because naming conventions for the states are non-uniform this class allows
-    you to enter the values at initialization.
-
-    Parameters
-    ----------
-    prefix : str
-        Full PPS Stopper PV.
-
-    name : str
-        Alias for the stopper.
-
-    in_state : str, optional
-        String associatted with in enum value.
-
-    out_state : str, optional
-        String associatted with out enum value.
-    """
-
-    state = Cpt(EpicsSignalRO, '', string=True, kind='hinted')
-    # QIcon for UX
-    _icon = 'fa.times-circle'
-
-    def __init__(self, prefix, *, in_state='IN', out_state='OUT', **kwargs):
-        # Store state information
-        self.in_states = [in_state]
-        self.out_states = [out_state]
-        self.states_list = self.in_states + self.out_states
-        # Load InOutPositioner
-        super().__init__(prefix, **kwargs)
-
-    def check_value(self, state):
-        """PPSStopper can not be commanded via EPICS."""
-        raise PermissionError("PPSStopper can not be commanded via EPICS")
 
 
 class ValveBase(Device):
